back-end/requirements.txt
asgiref==3.2.10
bson==0.5.10
Django==3.0.8
geog==0.0.2
numpy==1.19.0
paho-mqtt==1.5.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
